Chat now works. NEED Someone to overhall this please. See Desc
I need someone to take a look at the session details part so tidying up the UI to work better and see to the TODO on line 51 of UI.lua please.

Also needs the session members bit to be finished and made working. will require setup with the server too.

Finally the server has been moved to a seperate repository now which can be found here: https://github.com/Starystars67/BeamNG-MP-Server