add ui apps check and remove radial menu check
I don't think the reason for checking if the radial menu and removing it exists anymore, at least in BeamNG.drive v0.26, and for some versions now where it is not a default base game UI app. (I think it used to break things?). The radial menu is baked into the in-game menu these days, even though it can still be added as a UI app, and neither of these seem to cause issues.

I have also added a way to check if a player has REMOVED any of the three BeamMP UI apps from the multiplayer UI apps layout and replace whatever is missing with its default position and size settings upon returning to the main menu after exiting a server, or whenever the MP mod is loaded.