Prevent sort on nil in getFavorites function
This commit adds a nil check for the 'favs' table before attempting to sort it in the getFavorites function. If 'favs' is nil, which occurs when the favorites file is missing, empty, or improperly formatted, the sort operation will be skipped, and an error will be logged. Additionally, 'favs' is initialized to an empty table to prevent downstream errors that expect 'favs' to be a table.

Fixes an issue I was facing. 
```
Fatal Lua Error
string "lua/ge/extensions/MPConfig.lua":186: bad argument #1 to 'sort' (table expected, got nil)
=================== Stack Traceback >> START >> (1) field C function 'sort' (2) Lua field 'getFavorites' at line 186 of chunk 'lua/ge/extensions/MPConfig.lua' Local variables: favs = nil
```